requirements.txt
# Dependências de produção - versões fixadas para produção
Flask>=3.0,<4.0
Flask-SQLAlchemy>=3.0,<4.0
Flask-Migrate>=4.0,<5.0
Flask-Limiter>=3.0,<4.0
Flask-Talisman>=1.1,<2.0
Flask-Cors>=4.0,<5.0
Flask-WTF>=1.1,<2.0
WTForms-SQLAlchemy>=0.2,<1.0
Flask-Login>=0.6,<1.0
Flask-JWT-Extended>=4.7,<5.0
SQLAlchemy>=2.0,<3.0
gunicorn==22.0.0
celery>=5.3,<6.0
requests>=2.31,<3.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
# Google API
google-api-python-client==2.138.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dateutil>=2.8,<3.0
pydantic>=2.0,<3.0
prometheus_client>=0.17,<1.0
psutil>=5.9,<6.0
flask-restx>=1.2,<2.0
marshmallow>=3.20,<4.0
flask-marshmallow>=0.15,<1.0
greenlet>=3.0.0
SQLAlchemy-Utils==0.41.1
redis>=4.0.0

# Dependências de desenvolvimento e teste
pytest>=7.4,<8.0
pytest-cov>=4.1,<5.0
black==25.1.0
click==8.2.1
iniconfig==2.1.0
isort==6.0.1
mypy==1.16.1
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
Pygments==2.19.1
pyright==1.1.402
ruff==0.11.13
typing_extensions==4.14.0
flake8>=6.0,<7.0
types-requests>=2.31,<3.0
types-python-dateutil>=2.8,<3.0
Werkzeug>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-JWT-Extended
redis>=4.0.0
